Code_review_old/requirements.txt
google-adk
google-generativeai
python-dotenv
pylint
flake8